requirements.txt
python-telegram-bot==21.3
reportlab==4.2.5
Pillow==10.4.0
